requirements.txt
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.9
pymongo==4.5.0
python-dotenv==1.0.0 